refactor: move root lock functions to root/lock.c
