+++ Makefile.am
@@ -70,6 +70,7 @@ watchman_SOURCES = \
 	clientmode.c \
 	main.c       \
 	root.c       \
+	root/lock.c       \
 	state.c      \
 	string.c     \
 	time.c
